requirements.txt
streamlit>=1.29.0
moviepy>=1.0.3
openai-whisper>=20230314
torch>=2.0.0
numpy
ffmpeg-python
transformers>=4.36.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0
